requirements.txt
python-dotenv
rich
markov-word-generator
unidecode
typing-extensions
python-dateutil
psycopg[binary]
sshtunnel